requirements.txt
beautifulsoup4==4.11.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
numpy==1.22.4
Pillow==9.1.1
py-ogp-parser==0.2.1
PyYAML==6.0
requests==2.27.1
soupsieve==2.3.2.post1
urllib3==1.26.9
